x86 kernel updater test prep

+++ scripts/x86image.sh
@@ -82,11 +82,16 @@ sudo mount -t vfat "${BOOT_PART}" /mnt/volumio/rootfs/boot
 cp scripts/x86config.sh /mnt/volumio/rootfs
 if [ ! -d platform-x86 ]; then
   echo "Platform files (packages) not available yet, getting them from the repo"
-  git clone http://github.com/volumio/platform-x86
+  git clone http://github.com/gkkpch/platform-x86
 fi
-cp platform-x86/packages/linux-image-*.deb /mnt/volumio/rootfs
-cp platform-x86/packages/linux-firmware-*.deb /mnt/volumio/rootfs
-cp platform-x86/Intel-e1000e-3.3.4/e1000e.ko /mnt/volumio/rootfs
+if [ -f platform-x86/packages/.next ]; then
+  cp platform-x86/packages/experimental/linux-image-*.deb /mnt/volumio/rootfs
+  cp platform-x86/packages/experimental/linux-firmware-*.deb /mnt/volumio/rootfs
+else
+  cp platform-x86/packages/linux-image-*.deb /mnt/volumio/rootfs
+  cp platform-x86/packages/linux-firmware-*.deb /mnt/volumio/rootfs
+fi
+#cp platform-x86/Intel-e1000e-3.3.4/e1000e.ko /mnt/volumio/rootfs
 
 cp volumio/splash/volumio.png /mnt/volumio/rootfs/boot
